Preparing for ProcedureCall refactor.
-- procedure_call.rb
- Split out the majority of the functions and moved them out to
  procedure_call_helper.rb to be contained within the module
  ORAProcedureCallHelper.

-- procedure_call_helpers.rb
- Introduced a new module called: ProcedureCallHelperProvider that
  returns the correct helper module (either ORAProcedureCallHelper or
  PGProcedureCallHelper) depending on the dialect of the connection.

-- procedure_helpers.rb
- Started to populate the PGProcedureHelper module's functionality. This
  involved filling in some of the details of the get_argument_metadata
  method. However, this will be put on hold for a while, first need to
  fall back to implementing some of the ProcedureCall code.
